Work on sign payment transaction
    * Simplified sign() interface
    * Bugfix to passing of dataLength to handler functions
    * Added signing comfort screen UX
    * Added dataLength correctness check
    * Commented about log10()
    * Minor improvement to unit tests
    * Handle Mina connection exception better in wallet command-line

+++ src/getAddress.c
@@ -111,7 +111,7 @@ UX_STEP_TIMEOUT(
 UX_FLOW(ux_processing_flow,
         &ux_processing_step);
 
-void handleGetAddress(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags, volatile unsigned int *tx)
+void handleGetAddress(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint32_t dataLength, volatile unsigned int *flags, volatile unsigned int *tx)
 {
     UNUSED(dataLength);
     UNUSED(p2);
